requirements.txt
alembic==1.4.2
aniso8601==8.0.0
app==0.0.1
bcrypt==3.1.7
cached-property==1.5.1
cffi==1.14.1
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
flask-mysql-connector==1.0.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.7.1
mysql-connector-python==8.0.20
numpy==1.19.0
pandas==1.0.5
protobuf==3.12.2
pycparser==2.20
PyJWT==1.4.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
wt==2.1.1
WTForms==2.3.1
